Added enemyPoolService and its population

+++ Assets/Scripts/Logic/Player/IHealth.cs
@@ -1,4 +1,4 @@
 ﻿public interface IHealth
 {
-    void GetDamage(int damage);
+    void GetDamage(float damage);
 }
